fix: command
